app/requirements.txt
aiofiles==23.2.1
fastapi==0.103.1
mmcv==2.0.1
mmdeploy_runtime==1.2.0
numpy==1.24.4
opencv_python==4.8.0.76
python-dotenv==1.0.0
Requests==2.31.0
supabase==1.0.4
uvicorn==0.23.2
python-multipart==0.0.6
firebase_admin==6.2.0
